Refactor: Rename gapi file
