+++ ui/src/app/utils/project-billing-info.ts
